requirements.txt
yfinance>=0.2.36
pandas>=2.0.0
numpy>=1.26.0
scipy>=1.11.0
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.18.0
pyyaml>=6.0
python-dotenv>=1.0.0
jupyter>=1.0.0
pytest>=7.4.0